+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/exp/StreamTransport.scala
@@ -3,7 +3,6 @@ package com.twitter.finagle.http.exp
 import com.twitter.finagle.Status
 import com.twitter.finagle.transport.Transport
 import com.twitter.util.{Future, Time}
-import java.net.SocketAddress
 
 /**
  * A multi-part object with a single read handle, and a future that is satisfied
@@ -40,8 +39,6 @@ private[finagle] abstract class StreamTransportProxy[In, Out](val self: Transpor
 
   def status: Status = self.status
   val onClose: Future[Throwable] = self.onClose
-  def localAddress: SocketAddress = self.localAddress
-  def remoteAddress: SocketAddress = self.remoteAddress
   def close(deadline: Time): Future[Unit] = self.close(deadline)
   def context: Context = self.context
 }